Rename layers_list file to layerlist
