+++ napari/components/layerlist.py
