+++ napari/components/viewer_model.py
@@ -3,7 +3,7 @@ from math import inf
 from itertools import zip_longest
 from xml.etree.ElementTree import Element, tostring
 from .dims import Dims
-from .layers_list import LayersList
+from .layerlist import LayersList
 from .. import layers
 from ..util.event import EmitterGroup, Event
 from ..util.theme import palettes
